Do the same thing with List as with SList
